fix: TestController 삭제(#7)

+++ src/main/java/com/example/accchat3/domain/chat/controller/ChatController.java
@@ -21,7 +21,7 @@ public class ChatController {
 
 
     @GetMapping("/{chatRoomId}")
-    public ResponseEntity<GetChatRoomResponseDto> getChatRoom(@PathVariable Long chatRoomId){
+    public ResponseEntity<GetChatRoomResponseDto> getChatRoom(@PathVariable Integer chatRoomId){
         ChatRoom chatRoom=chatRoomService.getChatRoomById(chatRoomId);
         return ResponseEntity.ok(chatRoomMapper.toDto(chatRoom));
     }
